+++ packages/happner-2/lib/mesh.js
@@ -618,7 +618,6 @@ Mesh.prototype.initialize = util.promisify(function (config, callback) {
 // Step2 of two step start (initialize({},function callback(){ start() }))
 Mesh.prototype.start = util.promisify(function (callback) {
   var _this = this;
-
   //eslint-disable-next-line
   if (!_this._mesh.initialized) return console.warn('missing initialize()');
 
@@ -633,7 +632,7 @@ Mesh.prototype.start = util.promisify(function (callback) {
     Object.keys(_this._mesh.calls.starting).forEach(function (name) {
       _this.log.warn("awaiting startMethod '%s'", name);
     });
-  }, 10 * 1000);
+  }, _this._mesh.config.waitingInterval || 30e3);
 
   _this.impatient = setTimeout(function () {
     Object.keys(_this._mesh.calls.starting).forEach(function (name) {
@@ -643,7 +642,7 @@ Mesh.prototype.start = util.promisify(function (callback) {
       kill: true,
       wait: 200,
     });
-  }, _this._mesh.config.startTimeout || 5 * 60 * 1000); // Default to 5 minutes for AWS purposes
+  }, _this._mesh.config.startTimeout || 5 * 60e3); // Default to 5 minutes for AWS purposes
 
   _this.__initComponents(function (error) {
     if (error) {