+++ packages/happner-2/RELEASES.md
@@ -926,6 +926,11 @@ SMC-1645: made error messages more consistent, encapsulating mesh.component.meth
   - TEN-102: loki snapshot  file redundancy
   - TEN-92: made all dependencies test dependencies
 
+12.0.7 2022-04-06
+-----------------
+  - TEN-103: set default timeout on component loading wait warning to 30secs
+  - TEN-104: errors are only logged on system failures
+